refactor: factor out common code in worker choice strategies
