+++ src/pools/selection-strategies/weighted-round-robin-worker-choice-strategy.ts
@@ -1,4 +1,3 @@
-import { cpus } from 'node:os'
 import type { IWorker } from '../worker'
 import type { IPool } from '../pool'
 import { DEFAULT_WORKER_CHOICE_STRATEGY_OPTIONS } from '../../utils'
@@ -100,16 +99,5 @@ export class WeightedRoundRobinWorkerChoiceStrategy<
       this.workerVirtualTaskRunTime = 0
     }
     return true
-  }
-
-  private computeDefaultWorkerWeight (): number {
-    let cpusCycleTimeWeight = 0
-    for (const cpu of cpus()) {
-      // CPU estimated cycle time
-      const numberOfDigits = cpu.speed.toString().length - 1
-      const cpuCycleTime = 1 / (cpu.speed / Math.pow(10, numberOfDigits))
-      cpusCycleTimeWeight += cpuCycleTime * Math.pow(10, numberOfDigits)
-    }
-    return Math.round(cpusCycleTimeWeight / cpus().length)
   }
 }
